Update LGG Model.py
Added in propagation along the pump tube and the barrel. Still needs some polishing.
I've managed a workaround for the weird oscillatory behaviour in my last commit - as this only appeared to occur when the velocity of the piston dipped into negative numbers (i.e. if the pressure from the detonation wasn't sufficient to pop the rupture disk)